pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "catma"
version = "0.1.0"
description = "Category-theoretic modeling tools"
readme = "README.md"
authors = [{name = "Catma Contributors"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
  "pyyaml",
  "graphviz",
  "fastapi",
  "uvicorn",
  "streamlit",
]

[tool.setuptools.packages.find]
where = ["."]
